requirements.txt
alembic==1.10.4
async-timeout==4.0.2
Authlib==1.2.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
prettytable==3.7.0
pycparser==2.21
pyecharts==2.0.3
python-dotenv==1.0.0
redis==4.5.5
requests==2.30.0
simplejson==3.19.1
six==1.16.0
spotipy==2.23.0
SQLAlchemy==2.0.13
typing_extensions==4.5.0
urllib3==2.0.2
wcwidth==0.2.6
Werkzeug==2.3.4
WTForms==3.0.1
